backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
pandas
psycopg[binary]
